Move frontpage plot and dimension/metric select to own file
